+++ src/main/java/io/github/moulberry/notenoughupdates/miscfeatures/MiningStuff.java
@@ -22,7 +22,6 @@ package io.github.moulberry.notenoughupdates.miscfeatures;
 import io.github.moulberry.notenoughupdates.NotEnoughUpdates;
 import io.github.moulberry.notenoughupdates.autosubscribe.NEUAutoSubscribe;
 import io.github.moulberry.notenoughupdates.core.util.render.TextRenderUtils;
-import io.github.moulberry.notenoughupdates.miscfeatures.world.GenericBlockHighlighter;
 import io.github.moulberry.notenoughupdates.overlays.MiningOverlay;
 import io.github.moulberry.notenoughupdates.util.SBInfo;
 import io.github.moulberry.notenoughupdates.util.SpecialColour;
@@ -79,7 +78,7 @@ public class MiningStuff {
 					if (existingBlock.getBlock() == Blocks.stone &&
 						existingBlock.getValue(BlockStone.VARIANT) == BlockStone.EnumType.DIORITE_SMOOTH)
 						return;
-					if (!canPlayerSeeNearBlocks(pos.getX(), pos.getY(), pos.getZ()))
+					if (!Utils.canPlayerSeeNearBlocks(pos.getX(), pos.getY(), pos.getZ()))
 						return;
 					BlockPos player = Minecraft.getMinecraft().thePlayer.getPosition();
