+++ src/main/java/io/github/moulberry/notenoughupdates/miscfeatures/world/GenericBlockHighlighter.java
@@ -20,6 +20,7 @@
 package io.github.moulberry.notenoughupdates.miscfeatures.world;
 
 import io.github.moulberry.notenoughupdates.core.util.render.RenderUtils;
+import io.github.moulberry.notenoughupdates.util.Utils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.util.BlockPos;
@@ -57,44 +58,7 @@ public abstract class GenericBlockHighlighter {
 	public void onTick(TickEvent.ClientTickEvent ev) {
 		if (ev.phase != TickEvent.Phase.END) return;
 		highlightedBlocks.removeIf(it -> !isValidHighlightSpot(it) ||
-			!canPlayerSeeNearBlocks(it.getX(), it.getY(), it.getZ()));
-	}
-
-	public boolean canPlayerSeeNearBlocks(double x, double y, double z) {
-		EntityPlayerSP p = Minecraft.getMinecraft().thePlayer;
-		if (p == null) return false;
-		World world = p.worldObj;
-		Vec3 playerPosition = new Vec3(p.posX, p.posY + p.eyeHeight, p.posZ);
-		BlockPos blockPos = new BlockPos(x, y, z);
-		MovingObjectPosition hitResult1 = rayTraceBlocks(world, playerPosition, x, y, z);
-		if (canSee(hitResult1, blockPos)) return true;
-		MovingObjectPosition hitResult2 = rayTraceBlocks(world, playerPosition, x + 1, y, z);
-		if (canSee(hitResult2, blockPos.add(1, 0, 1))) return true;
-		MovingObjectPosition hitResult3 = rayTraceBlocks(world, playerPosition, x + 1, y + 1, z);
-		if (canSee(hitResult3, blockPos.add(1, 1, 0))) return true;
-		MovingObjectPosition hitResult4 = rayTraceBlocks(world, playerPosition, x + 1, y + 1, z + 1);
-		if (canSee(hitResult4, blockPos.add(1, 1, 1))) return true;
-
-		MovingObjectPosition hitResult5 = rayTraceBlocks(world, playerPosition, x, y + 1, z + 1);
-		if (canSee(hitResult5, blockPos.add(0, 1, 1))) return true;
-		MovingObjectPosition hitResult6 = rayTraceBlocks(world, playerPosition, x, y + 1, z);
-		if (canSee(hitResult6, blockPos.add(0, 1, 0))) return true;
-		MovingObjectPosition hitResult7 = rayTraceBlocks(world, playerPosition, x + 1, y, z + 1);
-		if (canSee(hitResult7, blockPos.add(1, 0, 1))) return true;
-		MovingObjectPosition hitResult8 = rayTraceBlocks(world, playerPosition, x, y + 1, z);
-		if (canSee(hitResult8, blockPos.add(0, 1, 0))) return true;
-
-		return false;
-	}
-
-	private static boolean canSee(MovingObjectPosition hitResult, BlockPos bp) {
-		return hitResult == null
-			|| hitResult.typeOfHit != MovingObjectPosition.MovingObjectType.BLOCK
-			|| bp.equals(hitResult.getBlockPos());
-	}
-
-	private static MovingObjectPosition rayTraceBlocks(World world, Vec3 playerPosition, double x, double y, double z) {
-		return world.rayTraceBlocks(playerPosition, new Vec3(x, y, z), false, true, true);
+			!Utils.canPlayerSeeNearBlocks(it.getX(), it.getY(), it.getZ()));
 	}
 
 	@SubscribeEvent
@@ -110,7 +74,7 @@ public abstract class GenericBlockHighlighter {
 		BlockPos blockPos = new BlockPos(x, y, z);
 		boolean contains = highlightedBlocks.contains(blockPos);
 		if (!contains) {
-			boolean canSee = canPlayerSeeNearBlocks(blockPos.getX(), blockPos.getY(), blockPos.getZ());
+			boolean canSee = Utils.canPlayerSeeNearBlocks(blockPos.getX(), blockPos.getY(), blockPos.getZ());
 			if (isValidHighlightSpot(blockPos) && canSee) {
 				highlightedBlocks.add(blockPos);
 			}
