+++ src/main/java/io/github/moulberry/notenoughupdates/util/Utils.java
@@ -59,13 +59,17 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
 import net.minecraft.nbt.NBTTagString;
 import net.minecraft.network.play.client.C0DPacketCloseWindow;
+import net.minecraft.util.BlockPos;
 import net.minecraft.util.ChatComponentText;
 import net.minecraft.util.ChatStyle;
 import net.minecraft.util.EnumChatFormatting;
 import net.minecraft.util.IChatComponent;
 import net.minecraft.util.Matrix4f;
+import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.StatCollector;
+import net.minecraft.util.Vec3;
+import net.minecraft.world.World;
 import net.minecraftforge.fml.common.Loader;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -2486,5 +2490,42 @@ public class Utils {
 		}
 
 		return "discord.gg/" + misc.get("discordInvite").getAsString();
+	}
+
+	public static boolean canPlayerSeeNearBlocks(double x, double y, double z) {
+		EntityPlayerSP p = Minecraft.getMinecraft().thePlayer;
+		if (p == null) return false;
+		World world = p.worldObj;
+		Vec3 playerPosition = new Vec3(p.posX, p.posY + p.eyeHeight, p.posZ);
+		BlockPos blockPos = new BlockPos(x, y, z);
+		MovingObjectPosition hitResult1 = rayTraceBlocks(world, playerPosition, x, y, z);
+		if (canSee(hitResult1, blockPos)) return true;
+		MovingObjectPosition hitResult2 = rayTraceBlocks(world, playerPosition, x + 1, y, z);
+		if (canSee(hitResult2, blockPos.add(1, 0, 1))) return true;
+		MovingObjectPosition hitResult3 = rayTraceBlocks(world, playerPosition, x + 1, y + 1, z);
+		if (canSee(hitResult3, blockPos.add(1, 1, 0))) return true;
+		MovingObjectPosition hitResult4 = rayTraceBlocks(world, playerPosition, x + 1, y + 1, z + 1);
+		if (canSee(hitResult4, blockPos.add(1, 1, 1))) return true;
+
+		MovingObjectPosition hitResult5 = rayTraceBlocks(world, playerPosition, x, y + 1, z + 1);
+		if (canSee(hitResult5, blockPos.add(0, 1, 1))) return true;
+		MovingObjectPosition hitResult6 = rayTraceBlocks(world, playerPosition, x, y + 1, z);
+		if (canSee(hitResult6, blockPos.add(0, 1, 0))) return true;
+		MovingObjectPosition hitResult7 = rayTraceBlocks(world, playerPosition, x + 1, y, z + 1);
+		if (canSee(hitResult7, blockPos.add(1, 0, 1))) return true;
+		MovingObjectPosition hitResult8 = rayTraceBlocks(world, playerPosition, x, y + 1, z);
+		if (canSee(hitResult8, blockPos.add(0, 1, 0))) return true;
+
+		return false;
+	}
+
+	private static boolean canSee(MovingObjectPosition hitResult, BlockPos bp) {
+		return hitResult == null
+			|| hitResult.typeOfHit != MovingObjectPosition.MovingObjectType.BLOCK
+			|| bp.equals(hitResult.getBlockPos());
+	}
+
+	private static MovingObjectPosition rayTraceBlocks(World world, Vec3 playerPosition, double x, double y, double z) {
+		return world.rayTraceBlocks(playerPosition, new Vec3(x, y, z), false, true, true);
 	}
 }
